+++ apps/docs/src/app/core/component-docs/popover/examples/popover-dropdown/popover-dropdown-example.component.ts
@@ -1,5 +1,10 @@
 import { Component } from '@angular/core';
+import { RtlService } from '@fundamental-ngx/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
+const bottomStart = 'bottom-start';
+const bottomEnd = 'bottom-end';
 @Component({
   selector: 'fd-popover-dropdown-example',
   templateUrl: './popover-dropdown.component.html',
@@ -7,9 +12,18 @@ import { Component } from '@angular/core';
 })
 export class PopoverDropdownExampleComponent {
 
-    menu = [
-        'Option 1',
-        'Option 2',
-        'Option 3',
-    ];
+  dropDownPlacement$: Observable<string>;
+
+  menu = [
+    'Option 1',
+    'Option 2',
+    'Option 3',
+  ];
+
+  constructor(private _rtlService: RtlService) {
+    this.dropDownPlacement$ = _rtlService.rtl
+      .pipe(
+        map(isRtl => isRtl ? bottomEnd : bottomStart)
+      )
+  }
 }