+++ libs/core/src/lib/list/list-item.directive.ts
@@ -49,5 +49,4 @@ export class ListItemDirective {
             this.itemEl.nativeElement.click();
         }
     }
-
 }