+++ libs/core/src/lib/mega-menu/mega-menu.component.ts
@@ -1,4 +1,10 @@
-import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, ViewEncapsulation, Optional } from '@angular/core';
+import { Observable, of } from 'rxjs';
+import { RtlService } from '../utils/public_api';
+import { map } from 'rxjs/operators';
+
+const rtl = 'rtl';
+const ltr = 'ltr';
 /**
  *  Component represents mega menu element, which contains list with menu items, links, sublists, subitems and sublinks..
  *  ```html
@@ -27,4 +33,17 @@ import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class MegaMenuComponent {}
+export class MegaMenuComponent {
+    dir$: Observable<string>;
+
+    constructor(@Optional() private rtlService: RtlService) {
+        if (rtlService) {
+            this.dir$ = rtlService.rtl
+                .pipe(
+                    map(isRtl => isRtl ? rtl : ltr)
+                )
+        } else {
+            this.dir$ = of(ltr);
+        }
+    }
+}